// src/models/enums.ts
export enum REGULATORY_CATEGORY {
    SIMILAR = "SIMILAR",
    HERBAL = "FITOTERÁPICO",
    NEW = "NOVO",
    BIOLOGICAL = "BIOLÓGICO",
    SPECIFIC = "ESPECÍFICO",
    GENERIC = "GENÉRICO",
    DYNAMIC = "DINAMIZADO",
    RADIOPHARMACEUTICAL = "RADIOFÁRMACO",
    T_PRODUCT = "PRODUTO DE T",
    OUTRO = "OUTRO"
  }
  
  export enum REGISTRATION_STATUS {
    EXPIRED_CANCELLED = "CADUCO/CANCELADO",
    VALID = "VÁLIDO",
    ACTIVE = "ATIVO",
    DESCONHECIDO = "DESCONHECIDO"
  }
  